requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
mysqlclient==2.2.4
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1